# swagger-new.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IronLogic Fitness Training API",
    "description": "API for managing fitness training data including clients, activities, schedules, and benchmarks",
    "version": "1.0.0",
    "contact": {
      "name": "IronLogic API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/benchmark-templates": {
      "get": {
        "tags": ["Benchmarks"],
        "summary": "Get all benchmark templates",
        "description": "Retrieve a list of all benchmark templates",
        "responses": {
          "200": {
            "description": "List of benchmark templates",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BenchmarkTemplate"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          }
        }
      },
      "post": {
        "tags": ["Benchmarks"],
        "summary": "Create a new benchmark template",
        "description": "Create a new benchmark template",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "benchmarkType"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200,
                    "description": "Template name"
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 1000,
                    "description": "Template notes"
                  },
                  "benchmarkType": {
                    "type": "string",
                    "enum": ["Lift", "Other"],
                    "description": "Type of benchmark"
                  }
                }
              },
              "example": {
                "name": "Bench Press Max",
                "notes": "1RM bench press template",
                "benchmarkType": "Lift"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Benchmark template created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/BenchmarkTemplate"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
    "/benchmark-templates/{id}": {
      "get": {
        "tags": ["Benchmarks"],
        "summary": "Get benchmark template by ID",
        "description": "Get a specific benchmark template by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Benchmark template ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Benchmark template found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/BenchmarkTemplate"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Benchmark template not found"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      },
      "put": {
        "tags": ["Benchmarks"],
        "summary": "Update benchmark template",
        "description": "Update a benchmark template",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Benchmark template ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "benchmarkType"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 1000
                  },
                  "benchmarkType": {
                    "type": "string",
                    "enum": ["Lift", "Other"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Benchmark template updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/BenchmarkTemplate"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Benchmark template not found"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      },
      "delete": {
        "tags": ["Benchmarks"],
        "summary": "Delete benchmark template",
        "description": "Delete a benchmark template",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Benchmark template ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Benchmark template deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Benchmark template not found"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BenchmarkTemplate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Template ID"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Template name"
          },
          "notes": {
            "type": "string",
            "maxLength": 1000,
            "description": "Template notes"
          },
          "benchmarkType": {
            "type": "string",
            "enum": ["Lift", "Other"],
            "description": "Type of benchmark"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": ["name", "benchmarkType"]
      },
...
    }
  }
} 